Use custom text size calculation
`Utils.textSize` was causing problems when running `Label#update`
because the repaint triggered when restyling the _textSizeNode was
sometimes not completed by the time `getBoundingClientRect` was called
and thus we were getting back 0 width/height and loosing our labels on
screen.

This method is a little more dependable, albeit a bit slower.